rust-toolchain: Update nightly toolchain (1.95.0)
The previous nightly toolchain was from 2026-02-13, which had partial
support for some of the changes that ended up in Rust 1.95.0. This
updates the toolchain to the same day the 1.95.0 stable tag was
created.

Also fixes clippy issues reported after the update.

+++ patina_dxe_core/src/allocator/fixed_size_block_allocator.rs
@@ -969,7 +969,7 @@ mod tests {
                 let ranges: Vec<_> = fsb.get_memory_ranges().collect();
                 assert_eq!(ranges.len(), 1);
 
-                let expected_start = allocated_address as usize + size_of::<AllocatorListNode>();
+                let expected_start = allocated_address + size_of::<AllocatorListNode>();
                 let expected_end = expected_start + allocation_size - size_of::<AllocatorListNode>();
                 assert_eq!(ranges[0], expected_start..expected_end);
             });
